new css for appengine, jinja fixed
