+++ requirements.txt
@@ -1,0 +1,2 @@
+Flask==1.1.2
+google-cloud-datastore==2.1.0
