+++ templates/index.html
@@ -3,30 +3,24 @@
 
   <body>
       {% block content %}
-    <h1>PittBett</h1>
+    <h1 class=main_title>Welcome to PittBett</h1>
+    <h2 class=games_title>Today's Games: {{ newlist |length }}</h2> 
     <ul class="games">
+        {% for game in newlist %}
         <li class="game_block">
+            <div class="game_time">
+                <h4 class="game_block_time"> {{ timelist[loop.index0] }} </h4>
+            </div>
             <div class="team_and_button">
-                <p class="game_block_team"> {{ team1A }}</p>
+                <p class="game_block_team">{{ game.0 }}</p>
                 <button class="game_block_button">Bet 1</button>
             </div>
-            <p class="game_block_time"> {{ time1 }} </p>
             <div class="team_and_button">
-                <p class="game_block_team">{{ team1B }}</p>
+                <p class="game_block_team">{{ game.1 }}</p>
                 <button class="game_block_button">Bet 2</button>
             </div>
         </li>
-        <li class="game_block">
-            <div class="team_and_button">
-                <p class="game_block_team"> {{ team2A }}</p>
-                <button class="game_block_button">Bet 1</button>
-            </div>
-            <p class="game_block_time"> {{ time2 }} </p>
-            <div class="team_and_button">
-                <p class="game_block_team">{{ team2B }}</p>
-                <button class="game_block_button">Bet 2</button>
-            </div>
-        </li>
+        {% endfor %}
     </ul>
     {% endblock %}
   </body>
